make RenderNode as an abstract class
